chore: Move test helper functions to common.go
